Move db and dialog activities into packages
