+++ app/src/main/java/net/justdave/mcstatus/dialogs/PrivacyDialog.java
@@ -1,4 +1,4 @@
-package net.justdave.mcstatus;
+package net.justdave.mcstatus.dialogs;
 
 import android.app.Dialog;
 import android.content.Context;
@@ -8,6 +8,8 @@ import android.text.Html;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
+
+import net.justdave.mcstatus.R;
 
 import java.io.BufferedReader;
 import java.io.IOException;
